Revert "Comments (#30) (#32)"
This reverts commit ba10818b75ef109b6c88512629c97bef718d9c7c.
