+++ src/components/MishnaView/ExcerptsView.tsx
@@ -1,15 +1,42 @@
-import React from 'react';
+import React, { useState } from 'react';
+import makeStyles from '@mui/styles/makeStyles';
+import Accordion from '@mui/material/Accordion';
 import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import Typography from '@mui/material/Typography';
 import ExcerptView from './ExcerptView';
 import { excerptsMap } from '../../inc/excerptUtils';
 import { useTranslation } from 'react-i18next';
-import { ExcerptsAccordion } from './exercptsAccordion';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    overflow: 'hidden',
+    '&.MuiPaper-root.MuiAccordion-root': { backgroundColor: 'rgba(0, 0, 0, .03)' },
+    '&:not(:last-child)': {
+      borderBottom: 0,
+    },
+    minHeight: '3.5rem',
+  },
+  rootExpanded: {
+    overflow: 'hidden',
+    '&.MuiPaper-root.MuiAccordion-root': { backgroundColor: 'rgba(0, 0, 0, .03)' },
+    flexGrow: 1,
+    '& >  .MuiCollapse-root': {
+      height: '100% !important',
+      overflow: 'scroll',
+    },
+    '& > .MuiCollapse-root > .MuiCollapse-wrapper > .MuiCollapse-wrapperInner > div > .MuiAccordionDetails-root': {
+      marginBottom: '5rem',
+    },
+  },
+}));
 
 export default function ExcerptsView(props) {
+  const classes = useStyles();
   const { excerpts, expanded, type } = props;
+  const [excerptBox, setExcerptBox] = useState(false);
   const { t } = useTranslation();
+  const rootClass = excerptBox ? classes.rootExpanded : classes.root;
 
   const title = excerptsMap.get(type)?.title || '';
   const filteredList = excerpts?.filter((excerpt) => excerpt.type === type);
@@ -19,7 +46,14 @@ export default function ExcerptsView(props) {
   }
 
   return (
-    <ExcerptsAccordion>
+    <Accordion
+      className={rootClass}
+      square
+      expanded={excerptBox}
+      onChange={() => {
+        setExcerptBox(!excerptBox);
+      }}
+    >
       <AccordionSummary>
         <Typography>
           {t(title)} - {filteredList.length}
@@ -32,6 +66,6 @@ export default function ExcerptsView(props) {
           ))}
         </div>
       </AccordionDetails>
-    </ExcerptsAccordion>
+    </Accordion>
   );
 }
