+++ src/components/edit/EditMishna/ExcerptDialog.tsx
@@ -16,7 +16,6 @@ export enum EXCERPT_TYPE {
   BIBLIO = 'BIBLIO',
   EXPLANATORY = 'EXPLANATORY',
   DICTIONARY = 'DICTIONARY',
-  COMMENT = 'COMMENT',
 }
 
 const GROUP_NOSACH = [EXCERPT_TYPE.NOSACH, EXCERPT_TYPE.BIBLIO, EXCERPT_TYPE.EXPLANATORY, EXCERPT_TYPE.DICTIONARY];
