+++ src/i18n/i18n.ts
@@ -51,15 +51,6 @@ i18n
           Combined: 'משולב',
           "Manuscripts": "כתבי יד",
           'Light mode': 'מצב תצוגה',
-          'Public Comments': 'הערות ציבוריות',
-          'Personal Comments': 'הערות אישיות',
-          'Required field': 'שדה חובה',
-          'Comment content': 'תוכן ההערה',
-          title: 'כותרת',
-          'Comment type': 'סוג ההערה',
-          'Public comment': 'הערה ציבורית',
-          'Personal comment': 'הערה אישית',
-          'Line number': 'מספר שורה',
           'Add parallel': 'הוסף מקבילה',
           'Sync parallels': 'סנכרן מקורות מקבילים',
           'Demo': 'מדגים יכולות'
