+++ src/inc/textUtils.ts
@@ -32,13 +32,3 @@ export function getWordOccurence(inText: string, offset: number, word: string): 
 // console.log('found',found)
 // return -1;
 // }
-
-export const getFirstAndLastWordOfString = (text: string) =>{
-  const words = text.split(' ');
-  const firstWord = words[0];
-  const lastWord = words.at(-1);
-  if (lastWord === ''){
-    return [firstWord, words.at(-2)];
-  }
-  return [firstWord, lastWord];
-}
