+++ src/store/actions/navigationActions.ts
@@ -2,9 +2,7 @@ import { RootState } from '..';
 import PageService from '../../services/pageService';
 import SettingsService from '../../services/settingsServince';
 import { iMishna } from '../../types/types';
-import { UserGroup } from '../reducers/authReducer';
 import { tryAsyncWithLoadingState } from './actionHelpers';
-import { getPrivateComments } from './commentsActions';
 import { setManuscriptsForChapter } from './relatedActions';
 
 export const REQUEST_START = 'REQUEST_START';
@@ -97,12 +95,10 @@ export function getCurrentTractate() {
 }
 
 export function getMishna(tractate: string, chapter: string, mishna: string) {
-  return async function (dispatch, getState) {
+  return async function (dispatch) {
     let mishnaData = await tryAsyncWithLoadingState(dispatch, PageService.getMishna(tractate, chapter, mishna));
     if (mishnaData) {
       dispatch(setCurrentMishna(mishnaData));
-      const isAuthenticated = getState().authentication.userGroup !== UserGroup.Unauthenticated;
-      isAuthenticated && dispatch(getPrivateComments());
     }
   };
 }
